Speedup with batch multiprocessimg TvDB matchings It only takes 13ms per on my machine, but doing it 1000 times will take time

+++ Shoko.Server/Providers/TvDB/TvDBLinkingHelper.cs
@@ -16,10 +16,11 @@ namespace Shoko.Server
     {
         #region TvDB Matching
 
-        public static void GenerateTvDBEpisodeMatches(int animeID)
+        public static void GenerateTvDBEpisodeMatches(int animeID, bool skipMatchClearing = false)
         {
             // wipe old links except User Verified
-            RepoFactory.CrossRef_AniDB_TvDB_Episode.DeleteAllUnverifiedLinksForAnime(animeID);
+            if (!skipMatchClearing)
+                RepoFactory.CrossRef_AniDB_TvDB_Episode.DeleteAllUnverifiedLinksForAnime(animeID);
 
             var matches = GetTvDBEpisodeMatches(animeID);
